@@ -5,7 +5,7 @@
     "scripts": {
     "ng": "ng",
     "build": "npm run env -s && ng build --prod",
-    "start": "npm run env -s && ng serve --aot --proxy-config proxy.conf.js",
+    "start": "npm run env -s && ng serve --proxy-config proxy.conf.js",
     "serve:sw": "npm run build -s && npx http-server ./dist -p 4200",
     "lint": "ng lint && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
     "test": "npm run env -s && ng test",
@@ -16,37 +16,36 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.3",
-    "@angular/common": "^7.2.3",
-    "@angular/compiler": "^7.2.3",
-    "@angular/core": "^7.2.3",
-    "@angular/forms": "^7.2.3",
-    "@angular/http": "^7.2.3",
-    "@angular/platform-browser": "^7.2.3",
-    "@angular/platform-browser-dynamic": "^7.2.3",
-    "@angular/router": "^7.2.3",
+    "@angular/animations": "^8.1.0",
+    "@angular/common": "^8.1.0",
+    "@angular/compiler": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/forms": "^8.1.0",
+    "@angular/platform-browser": "^8.1.0",
+    "@angular/platform-browser-dynamic": "^8.1.0",
+    "@angular/router": "^8.1.0",
     "@ngx-translate/core": "^11.0.1",
-    "@angular/service-worker": "^7.2.3",
-    "@angular/cdk": "^7.0.1",
-    "@angular/material": "^7.0.1",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "material-design-icons-iconfont": "^4.0.5",
+    "@angular/service-worker": "^8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/material": "^8.0.2",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "material-design-icons-iconfont": "^5.0.1",
     "hammerjs": "^2.0.8",
-    "core-js": "^2.6.3",
-    "rxjs": "^6.4.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "~7.3.0",
-    "@angular/compiler-cli": "^7.2.3",
-    "@angular/language-service": "^7.2.3",
-    "@angular-devkit/build-angular": "^0.13.0",
+    "@angular/cli": "~8.1.0",
+    "@angular/compiler-cli": "^8.1.0",
+    "@angular/language-service": "^8.1.0",
+    "@angular-devkit/build-angular": "^0.801.0",
+    "@angularclass/hmr": "^2.1.3",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
-    "@ngx-rocket/scripts": "^3.0.0",
-    "@types/jasmine": "^3.3.8",
+    "@ngx-rocket/scripts": "^4.0.0",
+    "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^8.10.13",
-    "codelyzer": "^5.0.0",
+    "@types/node": "^10.9.0",
+    "codelyzer": "^5.1.0",
     "htmlhint": "^0.11.0",
     "https-proxy-agent": "^2.0.0",
     "jasmine-core": "~3.4.0",
@@ -59,13 +58,13 @@
     "karma-jasmine-html-reporter": "^1.4.0",
     "karma-junit-reporter": "^1.2.0",
     "protractor": "~5.4.0",
-    "puppeteer": "^1.12.0",
-    "stylelint": "~10.0.1",
+    "puppeteer": "^1.17.0",
+    "stylelint": "~10.1.0",
     "stylelint-config-recommended-scss": "~3.3.0",
     "stylelint-config-standard": "~18.3.0",
-    "stylelint-scss": "~3.6.0",
-    "ts-node": "~8.1.0",
-    "tslint": "~5.16.0",
-    "typescript": "~3.2.0"
+    "stylelint-scss": "~3.9.0",
+    "ts-node": "^8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.0"
   }
 }
