@@ -1,4 +1,4 @@
-import { browser } from 'protractor';
+import { browser, ExpectedConditions as until } from 'protractor';
 import { LoginPage } from './page-objects/login.po';
 import { AppSharedPage } from './page-objects/app-shared.po';
 import { ShellPage } from './page-objects/shell.po';
@@ -22,6 +22,7 @@ describe('when the app loads', () => {
     });
 
     it('should display the hello message', async () => {
+      await browser.wait(until.visibilityOf(shell.welcomeText), 5000, 'Element taking too long to appear');
       expect(await shell.getParagraphText()).toEqual('Hello world !');
     });
   });
