@@ -52,7 +52,7 @@ export class AppComponent implements OnInit, OnDestroy {
         untilDestroyed(this)
       )
       .subscribe(event => {
-        const title = event['title'];
+        const title = event.title;
         if (title) {
           this.titleService.setTitle(this.translateService.instant(title));
         }
