@@ -3,16 +3,9 @@ import { CommonModule } from '@angular/common';
 import { HTTP_INTERCEPTORS, HttpClient, HttpClientModule } from '@angular/common/http';
 import { RouteReuseStrategy, RouterModule } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
+
 import { RouteReusableStrategy } from './route-reusable-strategy';
-import { AuthenticationService } from './authentication/authentication.service';
-import { CredentialsService } from './authentication/credentials.service';
-import { AuthenticationGuard } from './authentication/authentication.guard';
-import { I18nService } from './i18n.service';
 import { HttpService } from './http/http.service';
-import { HttpCacheService } from './http/http-cache.service';
-import { ApiPrefixInterceptor } from './http/api-prefix.interceptor';
-import { ErrorHandlerInterceptor } from './http/error-handler.interceptor';
-import { CacheInterceptor } from './http/cache.interceptor';
 
 @NgModule({
   imports: [
@@ -22,14 +15,6 @@ import { CacheInterceptor } from './http/cache.interceptor';
     RouterModule
   ],
   providers: [
-    AuthenticationService,
-    CredentialsService,
-    AuthenticationGuard,
-    I18nService,
-    HttpCacheService,
-    ApiPrefixInterceptor,
-    ErrorHandlerInterceptor,
-    CacheInterceptor,
     {
       provide: HttpClient,
       useClass: HttpService
