@@ -7,7 +7,9 @@ import { environment } from '@env/environment';
 /**
  * Prefixes all requests not starting with `http[s]` with `environment.serverUrl`.
  */
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class ApiPrefixInterceptor implements HttpInterceptor {
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
