@@ -8,7 +8,9 @@ import { HttpCacheService } from './http-cache.service';
  * Caches HTTP requests.
  * Use ExtendedHttpClient fluent API to configure caching for each request.
  */
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class CacheInterceptor implements HttpInterceptor {
 
   private forceUpdate = false;
@@ -37,7 +39,7 @@ export class CacheInterceptor implements HttpInterceptor {
       const cachedData = this.forceUpdate ? null : this.httpCacheService.getCacheData(request.urlWithParams);
       if (cachedData !== null) {
         // Create new response to avoid side-effects
-        subscriber.next(new HttpResponse(cachedData as Object));
+        subscriber.next(new HttpResponse(cachedData as object));
         subscriber.complete();
       } else {
         next.handle(request)
