@@ -11,7 +11,9 @@ const log = new Logger('ErrorHandlerInterceptor');
 /**
  * Adds a default error handler to all requests.
  */
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class ErrorHandlerInterceptor implements HttpInterceptor {
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
