@@ -34,7 +34,7 @@ const untilDestroyedSymbol = Symbol('untilDestroyed');
  * }
  * ```
  */
-export function untilDestroyed(instance: Object, destroyMethodName: string = 'ngOnDestroy') {
+export function untilDestroyed(instance: object, destroyMethodName: string = 'ngOnDestroy') {
   return <T>(source: Observable<T>) => {
     const originalDestroy = instance[destroyMethodName];
     const hasDestroyFunction = typeof originalDestroy === 'function';
@@ -48,7 +48,7 @@ export function untilDestroyed(instance: Object, destroyMethodName: string = 'ng
     if (!instance[untilDestroyedSymbol]) {
       instance[untilDestroyedSymbol] = new Subject();
 
-      instance[destroyMethodName] = function () {
+      instance[destroyMethodName] = function() {
         if (hasDestroyFunction) {
           originalDestroy.apply(this, arguments);
         }
