@@ -8,10 +8,11 @@
 // This is useful for granularity you might need beyond just the environment.
 // Note that as usual, any environment variables you expose through it will end up in your
 // bundle, and you should not use it for any sensitive information like passwords or keys.
-import env from './.env';
+import { env } from './.env';
 
 export const environment = {
   production: false,
+  hmr: true,
   version: env.npm_package_version + '-dev',
   serverUrl: '/api',
   defaultLanguage: 'en-US',
